move sequence buffer into the module definition
